Performance and memory usage improvements to the cache:
1. Extracting a single table worth of data from the column key cache is no longer an O(n*c) operation (n is the number of tables in the schema). It's now O(c), where c is the number of columns in the requested table.

Note, however, that extracting imported and exported key information is still O(r) on the number of relationships in the schema. In my testing, this hasn't been a significant problem. For highly-connected databases, this could turn out to be significant. The problem is that we'd have to be able to index the cached information twice: once for the importing table name, and once for the exporting table name. This extra complexity doesn't seem warranted by my testing.

2. The cached schema-worth-of-information is now stored in a map that uses soft references for the values. This lets the cached data vanish magically when the VM is under memory pressure.