Created a TableModel that has entries based on a result set.
Also created a fancy JTable that can export selections and sort. Although this JTable was made originally for the result set model any TableModel can be given to the FancyExportableJTable and the search, and exporting will work on that model.

These are used in the new table factory that creates FancyExportableJTable with a ResultSetTableModel.

The two formatters are used in the FancyExportableJTable to export any kind of TableModel. At current they are fairly simple and may need to be expanded on later.