Refactored the way tables are reverse-engineered in the same way as has already been done for columns, FK relationships, and indexes. The goal is to support refresh at the table container level.
Currently, the test shows this works when there was already at least one table in the table container prior to refresh. It does not work when the table container was empty, because the presense of tables as children is what makes something believe it's a table container. So the corresponding test currently fails.