Updated the getSchemas() query in both the 2005 and 2008 wrappers. Now both getSchemas() queries match!
The primary functional difference with these queries vs. the previous ones is that they tend to hide things that can be used as schemas but usually aren't (such as the db_denydatareader role, for example). These oddball sometimes-schema things do show up if they have something in them, which is nice.