NEW - bug 1719: Where and alias editor panes shrink after user deletes text http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1719
This change adds a parameter to the EditablePStyledText that defines
a minimum number of characters the text field will have. If the field
has fewer than the defined number of characters white space will be 
appended to the field when it is not editable. This is done to give
the fields a minimum space to give people a larger area to click on.

Piccolo's PNodes have no way to define a minimum width and the
PStyledText will relayout and redefine the bounds of the components
if the bounds are attempted to be set. Defining a minimum number of
characters for a field allows the PStyledText to compute the layout
in its normal layout mode instead of moving code from Piccolo into 
the EditablePStyledText.

It may be helpful in the future to submit a patch that adds minimum
and maximum widths and heights to the PNode.