NEW - bug 1720: Editor panes not moving when the QueryPen is scrolled http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1720
This change adds a listener to the camera on the canvas the editor pane
is placed on. The listener will move the editor pane as the camera 
shifts to position the editor relative to the field it's editing. Since
the editor is placed on the canvas instead of a layer the camera is
viewing the normal scrolling and camera shift does not move the editor.