NEW - bug 2020: A warning before executing a query with a missing join in the querypen http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2020
Added a method to the query to check if joins are missing.

Moved the query properties panel to Wabit as it is unused in the
library part of the query pen and Wabit specific properties are being
placed on it now. If the query pen wants to show a properties panel
in the future we should split up the panel then and allow passing in
additional properties.