Initial commit of the annotation processor for generating persister helper classes. Annotations have been created to tag persistable SPObject classes, constructor, constructor parameters, accessor, and mutator methods. The AbstractSPObject and Grant classes have been annotated with these new annotations. The rest of the SPObjects still need to be annotated.
Also the PersistedWabitObject, WabitObjectProperty and RemovedWabitObject classes from Wabit has been refactored into the library as PersistedSPObject, PersistedSPOProperty and RemovedSPObject. Since this is going into a branch, the ones in Wabit have not been removed yet. Once the branch has been merged with the trunk, we should remove these 3 classes from Wabit and use the new ones in the library.