Added a method to the PersistedSPObjectTest that confirms all properties on SPObjects can be persisted. This changed some of the setup of tests to add objects to an object tree.
Fixed the SPClassVisitor by removing the reset method. The reset method was used to clear the getter and setter lists 
when a superclass was being iterated over as the superclass could contain the same methods as the original object being 
inspected. However, this caused all of the methods to be removed from the object we are actually trying to persist and 
made persister helpers that only contained superclass methods. Now the methods are checked if they exist in the setter 
or getter list before adding.

Includes a fix to the SQLTable so removing a column in the table and the events it fires removes the SQLIndex.column from its index all happens in the same transaction.

This change includes a fix to SQLIndex as the commit event was not being fired to the undo manager. The begin was fired 
from SQLIndex when removing a column, then the SQLIndex would be removed, making the undo manger un-listen to the 
SQLIndex, and then the commit would fire and not be sent to the undo manager.

One test is still failing but requires moderate changes to the annotations processor classes and this commit is big enough.