NEW - bug 2713: Dragging a column into the primary key of a self referencing table throws IOOBE http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2713
This change increases the coupling between the SQLTable, SQLColumn, SQLRelationship and SQLIndex. Previously the interaction between these classes was done with listeners. The use of listeners introduced brittle code as the order of events mattered in some situations. Now the classes know how they relate to each other more directly and they call methods to properly update their components. This change also removes some odd behaviour previously in these classes like normalizePrimaryKey, the primaryKeySeq in SQLColumn, and the primary key boolean in SQLIndex.

The persistence layer was also fixed in this commit as the table taking a final index field used a previously untested part of the persisters. The persisters have been updated to handle adding a child to an object whose properties exist in the persist calls.