This is the major work of moving the events of populate to the foreground thread. This contains: -The populate methods add objects on the foreground thread for all SQLObjects. -The fetch SQLColumns now get all of the columns in the entire database to speed up population. One call to populateColumns on a table will populate all of the columns on all of the tables because it is faster in the long run. -The children population exceptions that were previously one per object is now a map so there can be a child populate exception for each child type under the SQLObject. This is done mainly for SQLTable's use.
Major outstanding issues:
-Threading concerns with populating columns of a table. The fetch columns takes a list of SQLTables. This is
retrieved by getting the parent's list of tables without populating. This is expected that the tables would
be populated previously but because they could be populated on the foreground thread while this runs in the
background thread the tables may only be partially populated or not populated at all.
-More testing is required to ensure all issues are addressed. Either more testing by hand or being able
to run multi-threaded tests is required, currently all of the background/foreground testing is being pushed
to the same thread in the unit tests.