NEW - bug 2840: "Persists do not follow after each other" error with spring layout http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2840
The SPPersisterListener now updates its list of properties to persist correctly when it receives a change from the server
while in a transaction. This is rare to occur but can if a transaction pushes changes to the foreground thread in parts
instead of one whole transaction. At current this is only done by the spring layout, which is now handled, and one case
of populate, which makes new objects.

The documentation has been updated on runInForeground to warn users of the method to be careful if they are pushing a 
transaction to the foreground in parts.