Changed the way check constraints and enumerations are added to SQLTypePhysicalProperties. There are now SQLCheckConstraint and SQLEnumeration classes that are children of SQLTypePhysicalProperties. This change allows multiple named constraints to be added. Any SQLObject class that contains SQLCheckConstraints should implement the SQLCheckConstraintContainer interface, which defines methods that retrieve lists of check constraints, adds check constraints, and removes check constraints.
Also, the foreign key column updater listener on SQLRelationship now listens to the entire hierarchy on its primary key SQLColumns. This means that changes to its UserDefinedSQLType, SQLTypePhysicalProperties and SQLCheckConstraints underneath the column will update the foreign key along with it. The change made in support of that is that SQLColumn no longer fires events on properties that do not belong to it. i.e. Change to nullability on SQLColumn should only delegate to its UserDefinedSQLType, and the event should be fired by UserDefinedSQLType only.