Added performance enhancements to the SPJSONMessageDecoder. A decode(JSONTokener) method has been created which has similar functionality to the already existing decode(JSONArray) method.
Normally when a JSONArray is created, it creates its own JSONTokener and parses all of the tokens. These tokens are saved in an ArrayList as a buffer. When you get a JSONObject at a particular index, it can return the JSONObject right away.

However, this process can cause memory usage to spike very high on extremely large JSONArrays. The new decode(JSONTokener) method will instead make the persister call immediately after parsing each JSONObject token. By doing so, the JSONObject token can be marked for garbage collection when it finishes the persister call.