Adding some extra controls for Designer.
When the persister listener is writing a whole object we can choose to have it skip certain child types of the object. This is useful for cloning only parts of an object.

The session persister can now be told to not disable magic when persisting changes. This is useful for using the persister in ways that should affect an object model which causes side effects.