Pulling the persist objects comparator out of the SPSessionPersister so we can provide extensions to the comparator.
This is done initially for Designer as the Taxonomy class needs to be loaded in a different pattern than the default comparator provides due to cyclical dependencies.

It may be useful in the future uses of the persister to improve the configuration of the comparator instead of extending the class.