Turns out the rollback of object moving and object creation was never implemented properly.
When an object is moved we add the pre-existing object back in but previously we were still adding in all of the properties to the rollback list. In addition we were trying to roll back values to unconditional creation changes which would roll back values, like boolean, to null, which would cause NPEs. This problem was attempted to be fixed for #138 but the change ended up causing additional problems where the rolled back object would try to remove final fields which would cause other errors.

This issue has been a problem since 2010 and was attempted to be fixed by the change to #138

test #1958